refactor: implement cqrs-es Query trait directly on EventBusHandle and remove redundant EventBusPublisher
