+++ agent_store/src/lib.rs
@@ -68,16 +68,14 @@ use shared_kernel::view_repository::DynViewRepository;
 use std::collections::HashMap;
 use std::sync::Arc;
 
-pub mod event_bus_publisher;
 pub mod event_source;
 pub mod event_verification;
 pub mod in_memory;
 pub mod mongodb;
 pub mod postgres;
 
-pub use event_bus_publisher::EventBusPublisher;
-use shared_kernel::event_bus::EventBusHandle;
 pub use event_source::MongoEventSource;
+use shared_kernel::event_bus::EventBusHandle;
 
 /// A generic command handler for a specific aggregate.
 ///
@@ -106,18 +104,8 @@ where
         &self,
         aggregate_id: &str,
         command: A::Command,
-        mut metadata: HashMap<String, String>,
+        metadata: HashMap<String, String>,
     ) -> Result<(), cqrs_es::AggregateError<A::Error>> {
-        metadata
-            .entry("occurred_at".to_string())
-            .or_insert_with(|| chrono::Utc::now().to_rfc3339());
-        metadata
-            .entry("correlation_id".to_string())
-            .or_insert_with(|| uuid::Uuid::new_v4().to_string());
-        metadata
-            .entry("causation_id".to_string())
-            .or_insert_with(|| uuid::Uuid::new_v4().to_string());
-
         self.cqrs.execute_with_metadata(aggregate_id, command, metadata).await
     }
 }
@@ -202,7 +190,7 @@ fn bus_publisher<A: Aggregate + 'static>(bus: &EventBusHandle) -> Vec<Box<dyn Qu
 where
     A::Event: serde::Serialize + DomainEvent,
 {
-    vec![Box::new(EventBusPublisher::<A>::new(bus.clone())) as Box<dyn Query<A>>]
+    vec![Box::new(bus.clone()) as Box<dyn Query<A>>]
 }
 
 pub async fn identity_state<CCB: CqrsComponentBuilder>(
@@ -461,5 +449,3 @@ pub async fn holder_state<CCB: CqrsComponentBuilder>(
         },
     }
 }
-
-
