+++ shared-kernel/Cargo.toml
@@ -26,7 +26,6 @@ uuid = { version = "1.4", features = ["v4", "fast-rng", "serde"] }
 workspace = true
 
 [dev-dependencies]
-
 serde = { version = "1.0", features = ["derive"] }
 
 [features]
